+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [1.2.0] - Other action function added
+* Now you can use the property otherActions to add an extra function to the tap or drag action
+
 ## [1.0.1] - Minor bug fixes
 * Minor bug fixes